+++ src/integrations/mixpanel/serializers/mixpanel-events.js
@@ -2,23 +2,43 @@ const moment = require('moment');
 const uuidV1 = require('uuid/v1');
 const JSONAPISerializer = require('jsonapi-serializer').Serializer;
 
-function MixpanelEventsSerializer(events, collectionName, meta) {
+function MixpanelEventsSerializer(events, collectionName, meta, options) {
   events = events.map(function (event) {
+    const specialPropertiesMap = {
+      $city: 'city',
+      $region: 'region',
+      $timezone: 'timezone',
+      $os: 'os',
+      $os_version: 'osVersion',
+      mp_country_code: 'country',
+      time: 'date'
+    };
+
+    Object.keys(event.properties).forEach(function (prop) {
+      if (specialPropertiesMap[prop]) {
+        event[specialPropertiesMap[prop]] = event.properties[prop];
+      } else {
+        event[prop] = event.properties[prop];
+      }
+
+      delete event.properties[prop];
+    });
+
     event.id = uuidV1();
-    event.date = moment(parseInt(event.properties.time, 10) * 1000).format('');
-    event.city = event.properties.$city;
-    event.os = event.properties.$os;
-    event.browser = event.properties.$browser;
-    event.browserVersion = event.properties.$browser_version;
+    event.date = moment(parseInt(event.date, 10) * 1000).format('');
 
     return event;
   });
 
-  const type = `${collectionName}_mixpanel_events`;
+  const type = collectionName + '_mixpanel_events';
+  const attrs = ['id', 'event', 'date', 'city', 'region', 'country', 'os', 'osVersion', 'browser'];
+
+  if (options.integrations.mixpanel.customProperties) {
+    attrs.push.apply(attrs, options.integrations.mixpanel.customProperties);
+  }
 
   return new JSONAPISerializer(type, events, {
-    attributes: ['id', 'event', 'date', 'city', 'os', 'browser',
-      'browserVersion'],
+    attributes: attrs,
     keyForAttribute: function (key) { return key; },
     meta,
   });